requirements.txt
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cryptography==36.0.0
discord.py==1.7.3
docopt==0.6.2
frozenlist==1.3.0
h11==0.12.0
idna==3.3
multidict==5.2.0
outcome==1.1.0
pipreqs==0.4.11
pycparser==2.21
PyNaCl==1.4.0
pyOpenSSL==21.0.0
requests==2.26.0
schedule==1.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.0.1
urllib3==1.26.7
wsproto==1.0.0
yarg==0.1.9